requirements.txt
absl-py==0.7.1
aiocontextvars==0.2.1
aiohttp==3.5.4
aiohttp-cors==0.7.0
alembic==1.0.10
alog==1.0.0
apache-beam==2.12.0
apex==0.1
argh==0.26.2
asn1crypto==0.24.0
astor==0.8.0
async-timeout==3.0.1
atari-py==0.1.15
atomicwrites==1.3.0
attrs==18.2.0
autobahn==19.1.1
Automat==0.7.0
avro-python3==1.9.0
backcall==0.1.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.7.1
bitmex-websocket==0.2.55
bleach==3.1.0
blis==0.2.4
bokeh==1.2.0
boto3==1.9.183
botocore==1.12.183
Bottleneck==1.2.1
bs4==0.0.1
cachetools==3.1.0
certifi==2019.6.16
cffi==1.11.5
chardet==3.0.4
Click==7.0
cliff==2.14.1
cloudpickle==1.2.1
cmd2==0.9.12
colorama==0.4.1
colorclass==2.2.0
colorlog==4.0.2
constantly==15.1.0
contextvars==2.3
convertdate==2.1.3
crcmod==1.7
croniter==0.3.29
cryptography==2.4.2
cvxopt==1.2.2
cycler==0.10.0
cymem==2.0.2
Cython==0.29.12
cytoolz==0.9.0.1
dask==2.0.0
dataclasses==0.6
dateparser==0.7.0
decorator==4.4.0
defusedxml==0.5.0
dill==0.2.9
distributed==2.0.1
docker-pycreds==0.4.0
docutils==0.14
elasticsearch==7.0.2
empyrical==0.5.0
entrypoints==0.3
ephem==3.7.6.0
fast-bert==0.1.2
fastai==1.0.54
fastavro==0.21.17
fasteners==0.15
fastprogress==0.1.21
faust==1.7.0
fbprophet==0.5
feedparser==5.2.1
Flask==1.1.1
flatten-json==0.1.7
freetype-py==2.1.0.post1
future==0.17.1
gast==0.2.2
gevent==1.4.0
gitdb2==2.0.5
GitPython==2.1.11
glob2==0.7
google-api-core==1.7.0
google-apitools==0.5.28
google-auth==1.6.2
google-auth-oauthlib==0.3.0
google-cloud-bigquery==1.9.0
google-cloud-bigtable==0.33.0
google-cloud-core==0.29.1
google-cloud-datastore==1.8.0
google-cloud-error-reporting==0.31.0
google-cloud-logging==1.10.0
google-cloud-monitoring==0.31.0
google-cloud-pubsub==0.39.1
google-pasta==0.1.7
google-resumable-media==0.3.2
googleapis-common-protos==1.5.6
googledatastore==7.1.0
gql==0.1.0
graphql-core==2.2
greenlet==0.4.15
grequests==0.3.0
grpc-google-iam-v1==0.11.4
grpcio==1.20.1
gym==0.12.5
h5py==2.9.0
hdfs==2.5.2
heapdict==1.0.0
holidays==0.9.10
httplib2==0.12.0
hyperlink==18.0.0
idna==2.8
idna-ssl==1.1.0
immutables==0.6
importlib-metadata==0.18
incremental==17.5.0
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kafka==1.3.5
kafka-python==1.4.6
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.0.1
lazyarray==0.3.3
llvmlite==0.29.0
locket==0.2.0
lomond==0.3.3
lunardate==0.2.0
lxml==4.3.4
Mako==1.0.12
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.2
mistune==0.8.4
mkl-fft==1.0.13
mkl-random==1.0.2
mock==2.0.0
mode==4.0.0
monotonic==1.5
more-itertools==7.0.0
mpi4py==2.0.0
msgpack==0.6.1
multidict==4.5.2
murmurhash==1.0.2
mypy-extensions==0.4.1
nb-conda==2.2.1
nb-conda-kernels==2.2.2
nbconvert==5.5.0
nbformat==4.4.0
neo==0.7.1
networkx==1.8.1
nltk==3.4.4
notebook==5.7.8
numba==0.44.1
numexpr==2.6.9
numpy==1.16.4
nvidia-ml-py3==7.352.0
oauth2client==4.1.3
oauthlib==3.0.1
okex==0.1.6
olefile==0.46
opencv-python==4.1.0.25
opentracing==1.3.0
optuna==0.12.0
packaging==19.0
pandas==0.24.2
pandas-datareader==0.7.0
pandocfilters==1.4.2
parsedatetime==2.4
parso==0.4.0
partd==1.0.0
pathtools==0.1.2
patsy==0.5.1
pbr==5.2.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
plac==0.9.6
pluggy==0.12.0
preshed==2.0.1
prettytable==0.7.2
prometheus-client==0.7.0
promise==2.2.1
prompt-toolkit==2.0.9
proto-google-datastore-v1==1.5.0
protobuf==3.8.0
psutil==5.6.3
ptyprocess==0.6.0
py==1.8.0
pyarrow==0.14.0
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
pydot==1.2.4
pyee==6.0.0
pyglet==1.3.2
Pygments==2.4.2
PyHamcrest==1.9.0
PyNN==0.9.4
pyOpenSSL==18.0.0
pyparsing==2.3.0
pyperclip==1.7.0
PyQt5==5.13.0
PyQt5-sip==4.19.18
pyqtgraph==0.10.0
pyrsistent==0.15.2
PySocks==1.7.0
pyspikes==0.2.2
pystan==2.19.0.0
pytest==4.6.3
pytest-mock==1.10.4
python-binance==0.7.1
python-dateutil==2.7.5
python-editor==1.0.4
python-snappy==0.5.4
pytimeparse==1.1.8
pytorch-pretrained-bert==0.6.2
pytz==2018.9
PyYAML==5.1.1
pyzmq==18.0.1
qtconsole==4.5.1
quantities==0.12.3
redis==3.2.1
regex==2018.11.22
requests==2.21.0
requests-oauthlib==1.2.0
rethinkdb==2.4.2.post1
robinhood-aiokafka==1.0.3
rsa==4.0
Rx==1.6.1
s3transfer==0.2.1
schedule==0.6.0
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
semver==2.8.1
Send2Trash==1.5.0
sentry-sdk==0.9.5
service-identity==18.1.0
shortuuid==0.5.0
six==1.12.0
sklearn==0.0
smmap2==2.0.5
sortedcontainers==2.1.0
soupsieve==1.9.2
spacy==2.1.4
SQLAlchemy==1.3.4
srsly==0.0.7
stable-baselines==2.7.0
statsmodels==0.10.0rc2
stevedore==1.30.1
subprocess32==3.5.4
ta==0.4.5
tblib==1.4.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
terminado==0.8.2
terminaltables==3.1.0
testpath==0.4.2
textblob==0.15.3
tflearn==0.3.2
thinc==7.0.4
timeloop==1.0.2
toolz==0.9.0
torch==1.1.0
torchvision==0.3.0
tornado==6.0.2
tqdm==4.32.1
traitlets==4.3.2
tweepy==3.7.0
tweet-preprocessor==0.5.0
Twisted==18.9.0
txaio==18.8.1
typing==3.6.6
typing-extensions==3.7.2
tzlocal==1.5.1
urllib3==1.24.1
vaderSentiment==3.2.1
venusian==1.2.0
vispy==0.6.0
wandb==0.8.4
wasabi==0.2.2
watchdog==0.9.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.15.4
widgetsnbextension==3.4.2
wrapt==1.11.2
xarray==0.11.3
xgboost==0.90
yarl==1.3.0
zict==1.0.0
zipp==0.5.1
zmq==0.0.0
zope.interface==4.6.0
